+++ ajax/dropdownTrackingDeviceType.php
@@ -33,7 +33,7 @@
  * ---------------------------------------------------------------------
  */
 
-use Glpi\Http\Response;
+use Glpi\Exception\Http\BadRequestHttpException;
 
 /** @var array $CFG_GLPI */
 global $CFG_GLPI;
@@ -49,7 +49,7 @@ $itemtype = $_POST["itemtype"] ?? '';
 
 // Check for required params
 if (empty($itemtype)) {
-    Response::sendError(400, "Bad request: itemtype cannot be empty", Response::CONTENT_TYPE_TEXT_HTML);
+    throw new BadRequestHttpException("Bad request: itemtype cannot be empty");
 }
 
 // Check if itemtype is valid in the given context